chore(refactor): Refactor filter button to be reusable
