+++ src/_css/FilterComponent.scss
@@ -6,6 +6,7 @@
   flex-direction: row;
   color: $filter-font-color;
   font-weight: 200;
+  padding-top: 20px;
 }
 
 #funnel {
